feat: LinkSearch기능 추가
