+++ package.json
@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@reduxjs/toolkit": "^2.1.0",
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
@@ -12,10 +13,10 @@
     "@types/react-dom": "^18.2.18",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-router-dom": "^6.21.1",
-    "styled-components": "^6.1.6",
     "react-modal": "^3.16.1",
+    "react-router-dom": "^6.21.1",
     "react-scripts": "5.0.1",
+    "styled-components": "^6.1.6",
     "typescript": "^4.9.5",
     "web-vitals": "^2.1.4"
   },
