--- src/pages/Folder.js
@@ -1,13 +1,0 @@
-import FolderSection from "../components/folder/FolderSection";
-import Header from "../components/folder/Header";
-import Footer from "../components/common/Footer";
-
-export default function Folder() {
-  return (
-    <>
-      <Header />
-      <FolderSection />
-      <Footer />
-    </>
-  );
-}
