--- src/pages/Shared.js
@@ -1,14 +1,0 @@
-import React from "react";
-import Header from "../components/shared/Header";
-import ShareSection from "../components/shared/ShareSection";
-import Footer from "../components/common/Footer";
-
-export default function Shared() {
-  return (
-    <>
-      <Header />
-      <ShareSection />
-      <Footer />
-    </>
-  );
-}
