+++ src/types.tsx
@@ -5,6 +5,11 @@ export interface AllSee {
   name: "전체";
 }
 
+export interface LinkSearchInputProps {
+  searchTerm: string;
+  setSearchTerm: React.Dispatch<React.SetStateAction<string>>;
+}
+
 export interface FolderListProps {
   onSelectFolder: (folder: UserFolder | AllSee) => void;
   selectedFolder: UserFolder | AllSee;
@@ -15,7 +20,7 @@ export interface FolderNameProps {
 }
 
 export interface FolderContentCardProps {
-  selectedFolder: UserFolder | AllSee;
+  items: UserLink[] | undefined;
 }
 
 export interface CardProps {
